+++ src/core/client/systems/hud.ts
@@ -1,5 +1,6 @@
 import * as alt from 'alt-client';
 import * as native from 'natives';
+import { HUD_IDENTIFIER } from '../../shared/enums/hudIdentifiers';
 
 import { PLAYER_SYNCED_META } from '../../shared/enums/playerSynced';
 import { SYSTEM_EVENTS } from '../../shared/enums/system';
@@ -27,9 +28,15 @@ let hudElements: Array<IHud> = [
         align: 2,
         scale: 0.5,
         color: new alt.RGBA(255, 255, 255, 225),
-        callback: (pos: { x: number; y: number }) => {
+        callback: (self: IHud, pos: { x: number; y: number }) => {
             const value = alt.Player.local.meta.cash ? alt.Player.local.meta.cash : 0;
             const fixedValue = parseFloat(value.toFixed(0));
+
+            if (self.callbackReroute) {
+                self.callbackReroute(fixedValue);
+                return null;
+            }
+
             return `~g~$${fixedValue.toLocaleString()} ~w~CASH`;
         },
     },
@@ -43,9 +50,15 @@ let hudElements: Array<IHud> = [
         align: 2,
         scale: 0.5,
         color: new alt.RGBA(255, 255, 255, 225),
-        callback: (pos: { x: number; y: number }) => {
+        callback: (self: IHud, pos: { x: number; y: number }) => {
             const value = alt.Player.local.meta.bank ? alt.Player.local.meta.bank : 0;
             const fixedValue = parseFloat(value.toFixed(0));
+
+            if (self.callbackReroute) {
+                self.callbackReroute(fixedValue);
+                return null;
+            }
+
             return ` ~g~$${fixedValue.toLocaleString()} ~w~BANK`;
         },
     },
@@ -59,8 +72,15 @@ let hudElements: Array<IHud> = [
         align: 2,
         scale: 0.5,
         color: new alt.RGBA(255, 255, 255, 225),
-        callback: (pos: { x: number; y: number }) => {
+        callback: (self: IHud, pos: { x: number; y: number }) => {
             const food = alt.Player.local.meta.food;
+
+            if (self.callbackReroute) {
+                const actualFood = food !== undefined && food !== null ? food.toFixed(0) : 100;
+                self.callbackReroute(actualFood);
+                return null;
+            }
+
             return food !== undefined && food !== null ? `${food.toFixed(0)} FOOD` : `100 FOOD`;
         },
     },
@@ -74,8 +94,15 @@ let hudElements: Array<IHud> = [
         align: 2,
         scale: 0.5,
         color: new alt.RGBA(255, 255, 255, 225),
-        callback: (pos: { x: number; y: number }) => {
+        callback: (self: IHud, pos: { x: number; y: number }) => {
             const water = alt.Player.local.meta.water;
+
+            if (self.callbackReroute) {
+                const actualWater = water !== undefined && water !== null ? water.toFixed(0) : 100;
+                self.callbackReroute(actualWater);
+                return null;
+            }
+
             return water !== undefined && water !== null ? `${water.toFixed(0)} WATER` : `100 WATER`;
         },
     },
@@ -89,7 +116,7 @@ let hudElements: Array<IHud> = [
         align: 0,
         scale: 0.5,
         color: new alt.RGBA(255, 255, 255, 225),
-        callback: () => {
+        callback: (self: IHud) => {
             let text = '';
             const isMetric = native.getProfileSetting(227);
             const currentSpeed = native.getEntitySpeed(alt.Player.local.vehicle.scriptID);
@@ -107,10 +134,16 @@ let hudElements: Array<IHud> = [
             text += `~o~${LocaleController.get(LOCALE_KEYS.VEHICLE_FUEL)} ${fuel.toFixed(2)}% ~w~| `;
 
             // Door Locks
-            if (native.getVehicleDoorLockStatus(alt.Player.local.vehicle.scriptID) === 2) {
+            const isLocked = native.getVehicleDoorLockStatus(alt.Player.local.vehicle.scriptID) === 2;
+            if (isLocked) {
                 text += `~r~${LocaleController.get(LOCALE_KEYS.VEHICLE_LOCKED)}`;
             } else {
                 text += `~g~${LocaleController.get(LOCALE_KEYS.VEHICLE_UNLOCKED)}`;
+            }
+
+            if (self.callbackReroute) {
+                self.callbackReroute(isMetric, speedCalc, fuel, isLocked);
+                return null;
             }
 
             return text;
@@ -127,7 +160,12 @@ let hudElements: Array<IHud> = [
         align: 0,
         scale: 0.5,
         color: new alt.RGBA(255, 255, 255, 225),
-        callback: () => {
+        callback: (self: IHud) => {
+            if (self.callbackReroute) {
+                self.callbackReroute(World.getTimeAsString());
+                return null;
+            }
+
             return World.getTimeAsString();
         },
     },
@@ -141,7 +179,12 @@ let hudElements: Array<IHud> = [
         align: 0,
         scale: 0.6,
         color: new alt.RGBA(255, 255, 255, 225),
-        callback: () => {
+        callback: (self: IHud) => {
+            if (self.callbackReroute) {
+                self.callbackReroute(objective);
+                return null;
+            }
+
             return objective ? objective : '';
         },
     },
@@ -155,9 +198,14 @@ let hudElements: Array<IHud> = [
         align: 0,
         scale: 0.75,
         color: new alt.RGBA(255, 255, 255, 225),
-        callback: (pos: { x: number; y: number }, scale: number) => {
+        callback: (self: IHud, pos: { x: number; y: number }, scale: number) => {
             const value = alt.Player.local.getSyncedMeta(PLAYER_SYNCED_META.WANTED_LEVEL);
             const stars = value !== null ? value : 0;
+
+            if (self.callbackReroute) {
+                self.callbackReroute(stars);
+                return null;
+            }
 
             for (let i = 0; i < 5; i++) {
                 const newPos = {
@@ -235,6 +283,28 @@ export class HudSystem {
     }
 
     /**
+     * Used to Overwrite a Default HUD Element
+     * Great for re-routing to your own WebView
+     * @static
+     * @param {HUD_IDENTIFIER} identifier
+     * @param {(...args: any[]) => void} overwriteCallback
+     * @return {*}
+     * @memberof HudSystem
+     */
+    static overwriteCallback(identifier: HUD_IDENTIFIER, overwriteCallback: (...args: any[]) => void) {
+        isUpdating = true;
+
+        const index = hudElements.findIndex((x) => x.identifier === identifier);
+        if (index <= -1) {
+            isUpdating = false;
+            return;
+        }
+
+        hudElements[index].callbackReroute = overwriteCallback;
+        isUpdating = false;
+    }
+
+    /**
      * Render HUD Elements
      * @static
      * @return {*}
@@ -257,7 +327,7 @@ export class HudSystem {
                 continue;
             }
 
-            const value = hudElements[i].callback(element.position, element.scale);
+            const value = hudElements[i].callback(hudElements[i], element.position, element.scale);
             if (!value) {
                 continue;
             }